broaden remit of joostice - beyond 'journalism'
joostice may support activities that do not fit the definition of
journalism, but achieve the same end. for example, data-visualisation
projects or supportive infrastructure for journos such as an open
source anonymous web application framework that could be used to create
wikileaks-type software, secure collaboration tools for citizen
journos, or uncensorable newsdiff sites. wording it as 'activities that
shed light on matters of public interest' captures many more
possibilities. wording suggested by donnie maclurcan - on the advice
that broad objects are best, as these objects becomes a legal document.
many possible project  would not be defined as 'journalism' in a legal
sense, but may be as intergral to our implicit hopes for what joostice
can help to accomplish